Fixed the logic of processing the received data (2)
