+++ lib/data/interactor/place_interactor.dart
@@ -1,4 +1,4 @@
-import 'package:places/data/model/place.dart';
+import 'package:places/domain/place.dart';
 import 'package:places/data/repository/place_repository.dart';
 
 class PlaceInteractor {
