+++ lib/domain/place.dart
@@ -1,0 +1,19 @@
+class Place {
+  final int id;
+  final double? lat;
+  final double? lng;
+  final String name;
+  final List<String> urls;
+  final String placeType;
+  final String description;
+
+  Place({
+    required this.id,
+    required this.lat,
+    required this.lng,
+    required this.name,
+    required this.urls,
+    required this.placeType,
+    required this.description,
+  });
+}
