+++ lib/mocks.dart
@@ -1,4 +1,4 @@
-import 'package:places/data/model/place.dart';
+import 'package:places/domain/place.dart';
 
 final List<Place> mocks = [
   Place(
