+++ lib/ui/screens/filters_screen.dart
@@ -2,9 +2,9 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:places/data/model/filters.dart';
-import 'package:places/data/model/location.dart';
-import 'package:places/data/model/place.dart';
+import 'package:places/domain/filters.dart';
+import 'package:places/domain/location.dart';
+import 'package:places/domain/place.dart';
 import 'package:places/main.dart';
 import 'package:places/ui/screens/res/colors.dart';
 import 'package:places/ui/screens/res/constants.dart' as Constants;
