+++ lib/ui/screens/res/constants.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:places/data/model/location.dart';
+import 'package:places/domain/location.dart';
 
 const Location userLocation = Location(56.8490809, 53.247786);
 const RangeValues defaultDistanceRange = RangeValues(0, 10000);
