+++ lib/ui/widgets/list_screen/card/sight_card.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:places/main.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:places/data/interactor/place_interactor.dart';
-import 'package:places/data/model/place.dart';
+import 'package:places/domain/place.dart';
 import 'package:places/data/repository/place_repository.dart';
 import 'package:places/mocks.dart';
 import 'package:places/ui/screens/res/icons.dart';
