+++ lib/ui/widgets/list_screen/search_bar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:places/data/model/filters.dart';
-import 'package:places/data/model/place.dart';
+import 'package:places/domain/filters.dart';
+import 'package:places/domain/place.dart';
 import 'package:places/ui/screens/filters_screen.dart';
 import 'package:places/ui/screens/res/colors.dart';
 import 'package:places/ui/screens/res/icons.dart';
