+++ lib/ui/widgets/list_screen/sliver_sights.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:places/data/model/place.dart';
+import 'package:places/domain/place.dart';
 import 'package:places/mocks.dart';
 import 'package:places/ui/widgets/list_screen/card/sight_card.dart';
 import 'package:places/ui/widgets/list_screen/card/sight_landscape_widget.dart';
