+++ lib/ui/widgets/visiting_screen/card/sight_visiting_landscape_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:places/data/model/place.dart';
+import 'package:places/domain/place.dart';
 import 'package:places/ui/widgets/visiting_screen/card/sight_card_favorite.dart';
 
 // Widget for displaying a list of places in horizontal orientation
